fix: get null status when api server is off

+++ src/services/setupInterceptors.js
@@ -5,6 +5,8 @@ import { refreshToken } from "../actions/auth";
 import { isValidToken } from "../functions";
 import { LOGOUT } from "../actions/types";
 
+const TOKEN_EXPIRED_MESSAGE =
+  "토큰이 만료되어 로그아웃되었습니다. 다시 로그인 해주세요.";
 /**
  * request, response를 받을때 error를 정의하는 interceptor
  */
@@ -30,19 +32,15 @@ const setup = (store) => {
     async (err) => {
       const originalConfig = err.config;
       if (originalConfig.url !== "/user/login" && err.response) {
-        // Access token was expired
-        // TODO: modify logic when refresh token === null
-
-        if (err.response.status === 401 && !originalConfig._retry) {
+        if (
+          err.response &&
+          err.response.status === 401 &&
+          !originalConfig._retry
+        ) {
           originalConfig._retry = true;
-          // FIXME: Fix Infinity loop after access token is expired
-          // when access token expires, client post refresh token if there is
-          // get refreshed access token, save to redux store and localstorage
           const rs = TokenService.getLocalRefreshToken();
           if (!isValidToken(rs)) {
-            window.alert(
-              "토큰이 만료되어 로그아웃되었습니다. 다시 로그인 해주세요."
-            );
+            window.alert(TOKEN_EXPIRED_MESSAGE);
             TokenService.removeUser();
             dispatch({ type: LOGOUT });
             return axiosInstance(originalConfig);
